+++ changelog.md
@@ -4,6 +4,12 @@ This file contains all the notable changes done to the Ballerina GraphQL package
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- [[#8201] Fix finite types are recurve types throwing errors](https://github.com/ballerina-platform/ballerina-library/issues/8201)
+
 ## [5.2.1] - 2025-08-18
 
 ### Fixed